fix: 分离 TheMovie 主键与 TMDB ID，修复多文件匹配同一 TMDB 电影时记录丢失
- TheMovie.ID 改为自增主键 (uint)，新增 TmdbId 字段存储 TMDB ID
- ChunkTheMovie 改为按 URL 查找更新，不再按 TMDB ID 覆盖
- Heart/Star/Played/PlayHistory 的 DataId 改为 uint 匹配新主键
- 修复所有因 ID 类型变更导致的编译错误 (service/controller/repository)
- 刮削失败的文件保留 PosterPath="/"，前端自动回退到默认/随机封面

+++ api/controllers/controllerErrFiles.go
@@ -325,7 +325,7 @@ func RefErrTheMovieById(c *gin.Context) {
 	}
 
 	// 只有当实际记录是新建的（ID 和 oldId 不同）时才迁移关联数据并删除旧记录
-	if actualMovie.ID != 0 && actualMovie.ID != oldId {
+	if actualMovie.ID != 0 && actualMovie.ID != uint(oldId) {
 		db.Model(&models.Star{}).Where("data_type = ? AND data_id = ?", "movie", oldId).Update("data_id", actualMovie.ID)
 		db.Model(&models.Played{}).Where("data_type = ? AND data_id = ?", "movie", oldId).Update("data_id", actualMovie.ID)
 		db.Model(&models.Heart{}).Where("data_type = ? AND data_id = ?", "movie", oldId).Update("data_id", actualMovie.ID)
